sensor class
