+++ app/src/main/java/com/github/wnder/guessLocation/GuessLocationCompass.java
@@ -1,0 +1,6 @@
+package com.github.wnder.guessLocation;
+
+public class GuessLocationCompass {
+
+
+}
