+++ app/src/main/res/layout/activity_guess_location.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".GuessLocationActivity"
+    tools:context=".guessLocation.GuessLocationActivity"
     android:id="@+id/guessLocationLayout">
 
     <com.mapbox.mapboxsdk.maps.MapView
